CLI webpack updated
